Fix disposals do_after

+++ Content.Server/GameObjects/Components/Disposal/DisposalMailingUnitComponent.cs
@@ -588,7 +588,7 @@ namespace Content.Server.GameObjects.Components.Disposal
                 "entryDelay",
                 0.5f,
                 seconds => _entryDelay = seconds,
-                () => (int) _entryDelay);
+                () => _entryDelay);
 
             serializer.DataField(ref _tag, "Tag", "");
         }
