+++ Content.Server/GameObjects/Components/Disposal/DisposalUnitComponent.cs
@@ -529,7 +529,7 @@ namespace Content.Server.GameObjects.Components.Disposal
                 "entryDelay",
                 0.5f,
                 seconds => _entryDelay = seconds,
-                () => (int) _entryDelay);
+                () => _entryDelay);
         }
 
         public override void Initialize()
